Tests: Move namespace_tests into task_test.py
Also remove global test clutter with uncleaned namespaces
